Resolution of problems with Entity Framework database-update and the Postings table. Changed connection string to TrustedConnection and removed the EF folder. Put the Posting info in the IdentityModels class.
