+++ Join Project/GuildBlog/GuildBlog.UI/GuildBlog.UI.csproj
@@ -192,18 +192,13 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Migrations\202004131444225_init.cs" />
-    <Compile Include="Migrations\202004131444225_init.designer.cs">
-      <DependentUpon>202004131444225_init.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Migrations\202004131741117_posting.cs" />
-    <Compile Include="Migrations\202004131741117_posting.designer.cs">
-      <DependentUpon>202004131741117_posting.cs</DependentUpon>
+    <Compile Include="Migrations\202004131817257_init.cs" />
+    <Compile Include="Migrations\202004131817257_init.designer.cs">
+      <DependentUpon>202004131817257_init.cs</DependentUpon>
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
-    <Compile Include="Models\EF\GuildBlogEntities.cs" />
-    <Compile Include="Models\EF\Posting.cs" />
+    <Compile Include="Models\Posting.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -384,11 +379,8 @@
     <Service Include="{4A0DDDB5-7A95-4FBF-97CC-616D07737A77}" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Migrations\202004131444225_init.resx">
-      <DependentUpon>202004131444225_init.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Migrations\202004131741117_posting.resx">
-      <DependentUpon>202004131741117_posting.cs</DependentUpon>
+    <EmbeddedResource Include="Migrations\202004131817257_init.resx">
+      <DependentUpon>202004131817257_init.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
