--- Join Project/GuildBlog/GuildBlog.UI/Models/EF/GuildBlogEntities.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.Entity;
-using System.Linq;
-using System.Web;
-
-namespace GuildBlog.UI.Models.EF
-{
-    public class GuildBlogEntities : DbContext
-    {
-        public GuildBlogEntities()
-            : base("GuildBlog")
-        {
-        }
-        public DbSet<Posting> Postings { get; set; }
-    }
-}
