+++ Join Project/GuildBlog/GuildBlog.UI/Web.config
@@ -9,7 +9,7 @@
   <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
   <connectionStrings>
-    <add name="DefaultConnection" connectionString="Server=localhost;Database=GuildBlog;User Id=BlogOwner;Password=testing123;" providerName="System.Data.SqlClient" />
+    <add name="DefaultConnection" connectionString="Server=localhost;Database=GuildBlog;Trusted_Connection=true;" providerName="System.Data.SqlClient" />
   </connectionStrings>
   <appSettings>
     <add key="webpages:Version" value="3.0.0.0" />
